stylesheet now detects if it is a media block
The stylesheet changes the 'in_media' status to
true when a media block is added. This must be turned
off but whatever code is adding nodes to stylesheet.

If the sheet has an active media block, then rulesets
are added to that instead of the stylesheet.

Stylesheet starts a new @media block automatically if another is
added immediatly after an exisitng one.

This approach ensures a tree representation of the heirarchy of
the original CSS. That is required to ensure certain optimisations
do not break the CSS.