can now merge 4 part longhand rules
This change converts 4 part longhands (margin-top, etc)
to shorthand (margin) if all 4 values exist.

If one of these is !important, then the new shorthand
rule will be important. This may not be what is wanted
but hopefully will be rare enough to be ignored.

(It could be fixed if it turns into an issue)